+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Added
+
+- `ServiceHost.get_metadata` for accessing per-request connection metadata from within an event handler.
+
 ## [0.1.0a13] - 2026-06-18
 
 ### Changed